Аудит кода, пакет 3: мелкие ошибки
Б5 заглавная кириллицы (HerbalistText.h) в строке ощущения и именах зелий;
Б6 Get Herbalist Grid не кэшируется; Б7 лагерь проверяет мир до сгорания
вязанки; Б9 защита данных реестров; П6 строка хранилища раз в 0.1 с, норма
биома один раз; мёртвые поля убраны; таймеры оберегов и время записи
Травника -- double. Тесты Herbalist.Text.*, эталон 771/0. Аудит закрыт.

+++ ProjectHerbalist/Source/ProjectHerbalist/Core/Journal/JournalTypes.h
@@ -80,8 +80,11 @@ struct PROJECTHERBALIST_API FJournalEntry
     // GetWorld()->GetTimeSeconds() — тот же единственный источник времени,
     // что уже используется для CreationTime предметов и суточного цикла
     // (AUDIT_AND_REFACTORING_PLAN §3.4, устранённое дублирование часов).
+    // double, как сами часы (аудит 2026-09-26): float терял доли секунды на
+    // долгом прохождении. Старые сейвы с float читаются -- тег свойства
+    // приводится к double при загрузке.
     UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = "Journal")
-    float GameTimeSeconds = 0.0f;
+    double GameTimeSeconds = 0.0;
 
     // Только для Type == MemoryFragment — само воспоминание (Def->TrueText/
     // FalseText, см. GridWorldManagerZaryana.cpp::CollectMemoryFragment).
